backend/requirements.txt
# Django Core
Django==5.0.9
djangorestframework==3.15.2
django-cors-headers==4.3.1

# Database (SQLite for development, PostgreSQL for production)
# psycopg2-binary==2.9.9  # Uncomment for PostgreSQL

# PDF Generation
WeasyPrint==62.3
Pillow==10.4.0

# Authentication & Security
supabase==2.8.0
gotrue==2.12.3
django-ratelimit==4.1.0

# Environment & Configuration
python-decouple==3.8
django-environ==0.11.2

# Template Engine
Jinja2==3.1.4

# Logging & Monitoring
django-extensions==3.2.3

# Testing
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1

# Development
black==24.8.0
flake8==7.1.1
isort==5.13.2

# Additional utilities
python-dateutil==2.9.0
requests==2.32.3